+++ sandbox/src/filters.rs
@@ -2,8 +2,8 @@ use slog::Logger;
 use warp::Filter;
 
 use crate::handlers::{
-    activate_protocol, bake_block_with_client, handle_rejection, init_client_data,
-    start_node_with_config, stop_node,
+    activate_protocol, bake_block_with_client, bake_block_with_client_arbitrary, get_wallets,
+    handle_rejection, init_client_data, start_node_with_config, stop_node,
 };
 use crate::node_runner::LightNodeRunnerRef;
 use crate::tezos_client_runner::{
@@ -24,8 +24,10 @@ pub fn sandbox(
     start(log.clone(), runner.clone())
         .or(stop(log.clone(), runner, client_runner.clone()))
         .or(init_client(log.clone(), client_runner.clone()))
+        .or(wallets(log.clone(), client_runner.clone()))
         .or(activate(log.clone(), client_runner.clone()))
-        .or(bake(log, client_runner))
+        .or(bake(log.clone(), client_runner.clone()))
+        .or(bake_random(log, client_runner))
         .recover(handle_rejection)
         .with(cors)
 }
@@ -67,6 +69,17 @@ pub fn init_client(
         .and_then(init_client_data)
 }
 
+pub fn wallets(
+    log: Logger,
+    client_runner: TezosClientRunnerRef,
+) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
+    warp::path!("wallets")
+        .and(warp::get())
+        .and(with_log(log))
+        .and(with_client_runner(client_runner))
+        .and_then(get_wallets)
+}
+
 pub fn activate(
     log: Logger,
     client_runner: TezosClientRunnerRef,
@@ -89,6 +102,17 @@ pub fn bake(
         .and(with_log(log))
         .and(with_client_runner(client_runner))
         .and_then(bake_block_with_client)
+}
+
+pub fn bake_random(
+    log: Logger,
+    client_runner: TezosClientRunnerRef,
+) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
+    warp::path!("bake")
+        .and(warp::get())
+        .and(with_log(log))
+        .and(with_client_runner(client_runner))
+        .and_then(bake_block_with_client_arbitrary)
 }
 
 fn json_body() -> impl Filter<Extract = (serde_json::Value,), Error = warp::Rejection> + Clone {