+++ sandbox/src/node_runner.rs
@@ -1,14 +1,15 @@
 use std::path::PathBuf;
-use std::process::{Child, Command};
+use std::process::{Child, Command, Stdio};
 use std::sync::{Arc, RwLock};
 use std::thread;
-use std::time::Duration;
+use std::{io::{Read, BufReader}, time::Duration};
 
 use failure::Fail;
 use nix::sys::signal::{self, Signal};
 use nix::unistd::Pid;
 use wait_timeout::ChildExt;
 use warp::reject;
+use itertools::Itertools;
 
 #[derive(Debug, Fail)]
 pub enum LightNodeRunnerError {
@@ -30,7 +31,7 @@ pub enum LightNodeRunnerError {
 
     /// Startup Error
     #[fail(display = "Error after light-node process spawned")]
-    NodeStartupError,
+    NodeStartupError { reason: String } ,
 }
 
 impl From<std::io::Error> for LightNodeRunnerError {
@@ -69,28 +70,54 @@ impl LightNodeRunner {
         }
     }
 
-    /// Spawn a light-node child process
+    /// Spawn a light-node child process with health check
     pub fn spawn(&mut self, cfg: serde_json::Value) -> Result<(), reject::Rejection> {
         if self.is_running() {
             Err(LightNodeRunnerError::NodeAlreadyRunning.into())
         } else {
+            // spawn the process for a "health check" with piped stderr
             let mut process = Command::new(&self.executable_path)
-                .args(Self::construct_args(cfg)?)
+                .args(Self::construct_args(cfg.clone())?)
+                .stderr(Stdio::piped())
                 .spawn()
-                .map_err(|err| reject::custom(LightNodeRunnerError::IOError { reason: err }))?;
+                .map_err(|err| LightNodeRunnerError::IOError { reason: err })?;
 
+            // give some time for the process to spawn, a second is sufficient, we only care if it fails to start up
+            // and if it does, it fails almost instatnly
             thread::sleep(Duration::from_secs(1));
 
+            // check for live process
             match process.try_wait() {
+                // process exited, we need to handle and show the exact error
                 Ok(Some(_)) => {
-                    //let error_msg = handle_stderr(&mut process);
-                    return Err(LightNodeRunnerError::NodeStartupError.into());
+                    let error_msg = handle_stderr(&mut process);
+                    return Err(LightNodeRunnerError::NodeStartupError{ reason: error_msg.split("USAGE:").take(1).join("").replace("error:", "").trim().into()}.into());
                 }
-                _ => (/* Do nothing*/),
-            }
+                // the process started up OK, but we restart it to enable normal logging (stderr won't be piped)
+                _ => {
+                    match Self::send_sigint(process.id()) {
+                        Ok(()) => {
+                            self.process = None;
+                        }
+                        // if for some reason, the SIGINT fails to end the process, kill it with SIGKILL
+                        Err(_) => {
+                            Self::terminate_ref(&mut process);
+                        }
+                    }
 
-            self.process = Some(process);
-            Ok(())
+                    // enable a to shut down gracefully
+                    thread::sleep(Duration::from_secs(4));
+
+                    // start the process again
+                    let process = Command::new(&self.executable_path)
+                        .args(Self::construct_args(cfg)?)
+                        .spawn()
+                        .map_err(|err| LightNodeRunnerError::IOError { reason: err })?;
+
+                    self.process = Some(process);
+                    Ok(())
+                }
+            }
         }
     }
 
@@ -99,7 +126,7 @@ impl LightNodeRunner {
         if self.is_running() {
             let process = self.process.as_mut().unwrap();
             // kill with SIGINT (ctr-c)
-            match signal::kill(Pid::from_raw(process.id() as i32), Signal::SIGINT) {
+            match Self::send_sigint(process.id()) {
                 Ok(()) => {
                     self.process = None;
                     Ok(())
@@ -155,4 +182,19 @@ impl LightNodeRunner {
             Err(LightNodeRunnerError::JsonParsingError.into())
         }
     }
+
+    /// Send SIGINT signal to the process with PID, light-node is cheking for this signal and shuts down
+    /// gracefully if recieved
+    fn send_sigint(pid: u32) -> Result<(), nix::Error>{
+        signal::kill(Pid::from_raw(pid as i32), Signal::SIGINT)
+    }
+}
+
+/// extract data as String form the piped stderr
+fn handle_stderr(process: &mut Child) -> String {
+    let stderr = process.stderr.take().unwrap();
+    let mut extract = String::new();
+    let _ = BufReader::new(stderr).read_to_string(&mut extract);
+
+    extract
 }