+++ sandbox/src/tezos_client_runner.rs
@@ -1,15 +1,19 @@
 use std::collections::HashMap;
 use std::fs;
-use std::io::Write;
+use std::io::{Write, BufReader, Read};
 use std::path::PathBuf;
 use std::process::Command;
 use std::sync::{Arc, RwLock};
 
 use failure::Fail;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 use warp::reject;
+use warp::http::StatusCode;
+use slog::{info, warn, Logger};
+use itertools::Itertools;
 
 use super::TEZOS_CLIENT_DIR;
+use crate::handlers::ErrorMessage;
 
 #[derive(Debug, Fail)]
 pub enum TezosClientRunnerError {
@@ -32,6 +36,10 @@ pub enum TezosClientRunnerError {
     /// Serde Error.
     #[fail(display = "Error in serde")]
     SerdeError { reason: serde_json::Error },
+
+    /// Call Error.
+    #[fail(display = "Tezos-client call error")]
+    CallError { message: ErrorMessage },
 }
 
 impl From<std::io::Error> for TezosClientRunnerError {
@@ -58,7 +66,7 @@ impl reject::Reject for TezosClientRunnerError {}
 pub type SandboxWallets = Vec<Wallet>;
 
 /// Structure holding data used by tezos client
-#[derive(Clone, Debug, Deserialize)]
+#[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Wallet {
     alias: String,
     public_key_hash: String,
@@ -71,6 +79,12 @@ pub struct Wallet {
 #[derive(Clone, Debug, Deserialize)]
 pub struct BakeRequest {
     alias: String,
+}
+
+#[derive(Serialize)]
+pub struct TezosClientErrorReply {
+    pub message: String,
+    pub field_name: String,
 }
 
 /// Thread-safe reference to the client runner
@@ -93,6 +107,27 @@ pub struct TezosProtcolActivationParameters {
     protocol_parameters: serde_json::Value,
 }
 
+#[derive(Serialize, Clone)]
+pub struct TezosClientReply {
+    pub output: String,
+    pub error: String,
+}
+
+impl TezosClientReply {
+    pub fn new(output: String, error: String) -> Self {
+        Self {
+            output,
+            error,
+        }
+    }
+}
+
+impl Default for TezosClientReply {
+    fn default() -> TezosClientReply {
+        TezosClientReply::new(String::new(), String::new())
+    }
+}
+
 impl TezosClientRunner {
     pub fn new(name: &str, executable_path: PathBuf, base_dir_path: PathBuf) -> Self {
         Self {
@@ -107,7 +142,9 @@ impl TezosClientRunner {
     pub fn activate_protocol(
         &self,
         mut activation_parameters: TezosProtcolActivationParameters,
-    ) -> Result<(), reject::Rejection> {
+    ) -> Result<TezosClientReply, reject::Rejection> {
+        let mut client_output: TezosClientReply = Default::default();
+
         // create a temporary file, the tezos-client requires the parameters to be passed in a .json file
         let mut file = fs::File::create("protocol_parameters.json")
             .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?;
@@ -161,21 +198,33 @@ impl TezosClientRunner {
                 &activation_parameters.timestamp,
             ]
             .to_vec(),
+            &mut client_output,
         )?;
 
         // remove the file after activation
         fs::remove_file("./protocol_parameters.json")
             .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?;
 
-        Ok(())
+        Ok(client_output)
     }
 
     /// Bake a block with the bootstrap1 account
-    pub fn bake_block(&self, request: BakeRequest) -> Result<(), reject::Rejection> {
-        let alias = if let Some(wallet) = self.wallets.get(&request.alias) {
-            &wallet.alias
+    pub fn bake_block(&self, request: Option<BakeRequest>) -> Result<TezosClientReply, reject::Rejection> {
+        let mut client_output: TezosClientReply = Default::default();
+
+        let alias = if let Some(request) = request {
+            if let Some(wallet) = self.wallets.get(&request.alias) {
+                &wallet.alias
+            } else {
+                return Err(TezosClientRunnerError::NonexistantWallet.into());
+            }
         } else {
-            return Err(TezosClientRunnerError::NonexistantWallet.into());
+            // if there is no wallet provided in the request (GET) set the alias to be an arbitrary wallet
+            if let Some(wallet) = self.wallets.values().next() {
+                &wallet.alias
+            } else {
+                return Err(TezosClientRunnerError::NonexistantWallet.into());
+            }
         };
 
         self.run_client(
@@ -191,16 +240,19 @@ impl TezosClientRunner {
                 &alias,
             ]
             .to_vec(),
+            &mut client_output,
         )?;
 
-        Ok(())
+        Ok(client_output)
     }
 
     /// Initialize the accounts in the tezos-client
     pub fn init_client_data(
         &mut self,
         requested_wallets: SandboxWallets,
-    ) -> Result<(), reject::Rejection> {
+    ) -> Result<TezosClientReply, reject::Rejection> {
+        let mut client_output: TezosClientReply = Default::default();
+
         self.run_client(
             [
                 "--base-dir",
@@ -216,6 +268,7 @@ impl TezosClientRunner {
                 "unencrypted:edsk31vznjHSSpGExDMHYASz45VZqXN4DPxvsa4hAyY8dHM28cZzp6",
             ]
             .to_vec(),
+            &mut client_output,
         )?;
 
         for wallet in requested_wallets {
@@ -234,30 +287,80 @@ impl TezosClientRunner {
                     &format!("unencrypted:{}", &wallet.secret_key),
                 ]
                 .to_vec(),
+                &mut client_output,
             )?;
             self.wallets.insert(wallet.alias.clone(), wallet);
         }
 
-        Ok(())
+        Ok(client_output)
     }
 
     /// Cleanup the tezos-client directory
-    pub fn cleanup(&self) -> Result<(), reject::Rejection> {
+    pub fn cleanup(&mut self) -> Result<(), reject::Rejection> {
         fs::remove_dir_all(TEZOS_CLIENT_DIR)
             .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?;
         fs::create_dir(TEZOS_CLIENT_DIR)
             .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?;
 
+        self.wallets.clear();
+
         Ok(())
     }
 
     /// Private method to run the tezos-client as a subprocess and wait for its completion
-    fn run_client(&self, args: Vec<&str>) -> Result<(), reject::Rejection> {
-        let _ = Command::new(&self.executable_path)
+    fn run_client(&self, args: Vec<&str>, client_output: &mut TezosClientReply) -> Result<(), reject::Rejection> {
+        let output = Command::new(&self.executable_path)
             .args(args)
-            .spawn()
-            .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?
-            .wait();
+            .output()
+            .map_err(|err| reject::custom(TezosClientRunnerError::IOError { reason: err }))?;
+        let _ = BufReader::new(output.stdout.as_slice()).read_to_string(&mut client_output.output);
+        let _ = BufReader::new(output.stderr.as_slice()).read_to_string(&mut client_output.error);
+
         Ok(())
+    }
+}
+
+/// Construct a reply using the output from the tezos-client
+pub fn reply_with_client_output(reply: TezosClientReply, log: &Logger) -> Result<impl warp::Reply, reject::Rejection> {
+    if reply.error.is_empty() {
+        info!(log, "Successfull tezos-client call: {}", reply.output);
+        Ok(StatusCode::OK)
+    } else {
+        if reply.output.is_empty() {
+            // error
+            if let Some((field_name, message)) = extract_field_name_and_message_ocaml(&reply) {
+                Err(TezosClientRunnerError::CallError { message: ErrorMessage::validation(500, message, field_name)}.into())
+            } else {
+                // generic
+                warn!(log, "GENERIC ERROR in tezos-client, log: {}", reply.error);
+                Err(TezosClientRunnerError::CallError { message: ErrorMessage::generic(500, "Unexpexted error in tezos-client call".to_string())}.into())
+            }
+        } else {
+            // this is just a warning
+            warn!(log, "Succesfull call with a warning: {} -> WARNING: {}", reply.output, reply.error);
+            Ok(StatusCode::OK)
+        }
+    }
+}
+
+/// Parse the returned error string from the tezos client
+fn extract_field_name_and_message_ocaml(reply: &TezosClientReply) -> Option<(String, String)>{
+    let parsed_message = reply.error.replace("\\", "").split("\"").filter(|s| s.contains("Invalid protocol_parameters")).join("").replace(" n{ ", "").replace("{", "");
+
+    // extract the field name depending on the parsed error
+    let field_name = if parsed_message.contains("Missing object field") {
+        Some(parsed_message.split_whitespace().last().unwrap_or("").to_string())
+    } else if parsed_message.contains("/") {
+        Some(parsed_message.split_whitespace().filter(|s| s.contains("/")).join("").replace("/", "").replace(",", ""))
+    } else {
+        None
+    };
+
+    if let Some(field_name) = field_name {
+        // simply remove the field name from the error message
+        let message = parsed_message.replace(&field_name, "").replace("At /, ", "").trim().to_string();
+        Some((field_name, message))
+    } else {
+        None
     }
 }